remove llmqtype cast not needed
Former-commit-id: 09970ccc28c98fa12a68cd5af8d77add3f12a74b

+++ src/evo/deterministicmns.cpp
@@ -889,7 +889,7 @@ void CDeterministicMNManager::HandleQuorumCommitment(const llmq::CFinalCommitmen
 {
     // The commitment has already been validated at this point so it's safe to use members of it
 
-    auto members = llmq::CLLMQUtils::GetAllQuorumMembers((uint8_t)qc.llmqType, pindexQuorum);
+    auto members = llmq::CLLMQUtils::GetAllQuorumMembers(qc.llmqType, pindexQuorum);
 
     for (size_t i = 0; i < members.size(); i++) {
         if (!mnList.HasMN(members[i]->proTxHash)) {
